Rename repo and move gurobi package to subfolder
