+++ .gitignore
@@ -162,4 +162,4 @@ cython_debug/
 .vscode/
 
 # Version file - autogenerated by hatch-vcs
-src/cvxpy_gurobi/_version.py
+src/cvxpy_translation/_version.py
