+++ pyproject.toml
@@ -3,9 +3,9 @@ build-backend = "hatchling.build"
 requires = ["hatchling", "hatch-vcs"]
 
 [project]
-name = "cvxpy-gurobi"
+name = "cvxpy-translation"
 dynamic = ["version"]
-description = "Translate CVXPY problems into gurobipy models"
+description = "Translate CVXPY problems into solvers' native models"
 readme = "README.md"
 keywords = [
   "cvxpy",
@@ -35,19 +35,22 @@ classifiers = [
 dependencies = [
   # "cvxpy" must be available but we don't list it
   # as a dependency because cvxpy-base is sufficient
-  "gurobipy",
   "numpy>=1.23.0",
   "scipy>=1.4.0",
 ]
+
+[project.optional-dependencies]
+gurobi = ["gurobipy"]
+
 [project.urls]
-Documentation = "https://github.com/jonathanberthias/cvxpy-gurobi#readme"
-Issues = "https://github.com/jonathanberthias/cvxpy-gurobi/issues"
-Source = "https://github.com/jonathanberthias/cvxpy-gurobi"
+Documentation = "https://github.com/jonathanberthias/cvxpy-translation#readme"
+Issues = "https://github.com/jonathanberthias/cvxpy-translation/issues"
+Source = "https://github.com/jonathanberthias/cvxpy-translation"
 
 [tool.hatch]
 version.source = "vcs"
 version.raw-options = { local_scheme = "no-local-version" }
-build.hooks.vcs.version-file = "src/cvxpy_gurobi/_version.py"
+build.hooks.vcs.version-file = "src/cvxpy_translation/_version.py"
 
 [tool.hatch.envs.default]
 description = "Base environment for other envs to inherit from"
@@ -143,10 +146,10 @@ ignore = [
   "S101",    # asserts used for type narrowing
 ]
 exclude = [
-  "src/cvxpy_gurobi/_version.py", # autogenerated by hatch-vcs
+  "src/cvxpy_translation/_version.py", # autogenerated by hatch-vcs
 ]
 [tool.ruff.lint.extend-per-file-ignores]
-"src/cvxpy_gurobi/translation.py" = [
+"src/cvxpy_translation/gurobi/translation.py" = [
   "N802",    # visitor uses uppercase letters in method names
   "PLR6301", # method could be a function or staticmethod
   "PLR0904", # too many public methods in class
@@ -161,7 +164,7 @@ exclude = [
 
 [tool.ruff.lint.isort]
 force-single-line = true
-known-first-party = ["cvxpy_gurobi", "test_problems"]
+known-first-party = ["cvxpy_translation", "test_problems"]
 
 [tool.ruff.lint.pylint]
 max-args = 10
@@ -177,7 +180,7 @@ testpaths = ["tests"]
 xfail_strict = true
 
 [tool.coverage.run]
-source_pkgs = ["cvxpy_gurobi"]
+source_pkgs = ["cvxpy_translation"]
 branch = true
 dynamic_context = "test_function"
 omit = [
