--- src/cvxpy_gurobi/__init__.py
@@ -1,31 +1,0 @@
-from cvxpy_gurobi._version import __version__
-from cvxpy_gurobi._version import __version_tuple__
-from cvxpy_gurobi.interface import GUROBI_TRANSLATION
-from cvxpy_gurobi.interface import backfill_problem
-from cvxpy_gurobi.interface import build_model
-from cvxpy_gurobi.interface import register_solver
-from cvxpy_gurobi.interface import solve
-from cvxpy_gurobi.translation import InvalidNonlinearAtomError
-from cvxpy_gurobi.translation import InvalidNormError
-from cvxpy_gurobi.translation import InvalidParameterError
-from cvxpy_gurobi.translation import InvalidPowerError
-from cvxpy_gurobi.translation import UnsupportedConstraintError
-from cvxpy_gurobi.translation import UnsupportedError
-from cvxpy_gurobi.translation import UnsupportedExpressionError
-
-__all__ = (
-    "GUROBI_TRANSLATION",
-    "InvalidNonlinearAtomError",
-    "InvalidNormError",
-    "InvalidParameterError",
-    "InvalidPowerError",
-    "UnsupportedConstraintError",
-    "UnsupportedError",
-    "UnsupportedExpressionError",
-    "__version__",
-    "__version_tuple__",
-    "backfill_problem",
-    "build_model",
-    "register_solver",
-    "solve",
-)
