+++ src/cvxpy_translation/__init__.py
@@ -1,0 +1,4 @@
+from cvxpy_translation._version import __version__
+from cvxpy_translation._version import __version_tuple__
+
+__all__ = ("__version__", "__version_tuple__")
