+++ src/cvxpy_translation/gurobi/__init__.py
@@ -1,0 +1,27 @@
+from cvxpy_translation.gurobi.interface import GUROBI_TRANSLATION
+from cvxpy_translation.gurobi.interface import backfill_problem
+from cvxpy_translation.gurobi.interface import build_model
+from cvxpy_translation.gurobi.interface import register_solver
+from cvxpy_translation.gurobi.interface import solve
+from cvxpy_translation.gurobi.translation import InvalidNonlinearAtomError
+from cvxpy_translation.gurobi.translation import InvalidNormError
+from cvxpy_translation.gurobi.translation import InvalidParameterError
+from cvxpy_translation.gurobi.translation import InvalidPowerError
+from cvxpy_translation.gurobi.translation import UnsupportedConstraintError
+from cvxpy_translation.gurobi.translation import UnsupportedError
+from cvxpy_translation.gurobi.translation import UnsupportedExpressionError
+
+__all__ = (
+    "GUROBI_TRANSLATION",
+    "InvalidNonlinearAtomError",
+    "InvalidNormError",
+    "InvalidParameterError",
+    "InvalidPowerError",
+    "UnsupportedConstraintError",
+    "UnsupportedError",
+    "UnsupportedExpressionError",
+    "backfill_problem",
+    "build_model",
+    "register_solver",
+    "solve",
+)
