+++ tests/test_non_optimal.py
@@ -2,13 +2,13 @@ import cvxpy as cp
 import gurobipy as gp
 from cvxpy import settings as s
 
-import cvxpy_gurobi
-from cvxpy_gurobi.translation import CVXPY_VERSION
+import cvxpy_translation.gurobi
+from cvxpy_translation.gurobi.translation import CVXPY_VERSION
 
 
 def test_backfill_unbounded() -> None:
     problem = cp.Problem(cp.Maximize(cp.Variable()))
-    cvxpy_gurobi.solve(problem)
+    cvxpy_translation.gurobi.solve(problem)
     model = problem.solver_stats.extra_stats
     assert problem.status == s.UNBOUNDED
     if CVXPY_VERSION >= (1, 2, 0):
@@ -18,7 +18,7 @@ def test_backfill_unbounded() -> None:
 def test_backfill_infeasible() -> None:
     x = cp.Variable(nonneg=True)
     problem = cp.Problem(cp.Maximize(x), [x <= -1])
-    cvxpy_gurobi.solve(problem, **{gp.GRB.Param.DualReductions: 0})
+    cvxpy_translation.gurobi.solve(problem, **{gp.GRB.Param.DualReductions: 0})
     model = problem.solver_stats.extra_stats
     assert problem.status == s.INFEASIBLE
     if CVXPY_VERSION >= (1, 2, 0):
